option to not install/configure https
added option to not install/config https. For when a different cert provider is to be used or the user wants to configure certs manually.

+++ main.sh
@@ -2,6 +2,7 @@
 
 staging=false
 cf=false
+https=true
 
 user_check() {
 	if [ "$username" == "mastodon" ]; then
@@ -14,7 +15,7 @@ user_check() {
 ## Check if rsync is installed. Install if required (Debian 10/11 - not installed by default?)
 type -p rsync >/dev/null || apt-get install rsync -y
 
-while getopts ":c:e:d:u:s" flag; do
+while getopts ":c:e:d:u:sn" flag; do
     case ${flag} in
         c  ) cloudflare_token=${OPTARG}
              cf=true
@@ -27,6 +28,8 @@ while getopts ":c:e:d:u:s" flag; do
 	         user_check
 	         ;;
 	    s  ) staging=true
+             ;;
+        n  ) https=false
              ;;
     esac
 done
@@ -59,10 +62,6 @@ if [ -z "$email" ]; then
     read email
 fi
 
-if [ $(awk -F= '/^ID=/{print $2}' /etc/os-release) == "debian" ]; then
-    apt install snapd -y
-fi
-
 # This is messy. There's another check for Ubuntu 22.04 at the end, but still... find a better way.
 ur=0
 if [ $(awk -F= '/^ID=/{print $2}' /etc/os-release) == "ubuntu" ]; then
@@ -75,16 +74,21 @@ if [ "$ur" = "22.04" ]; then
     sed -i 's/#$nrconf{ucodehints} = 0;/$nrconf{ucodehints} = 0;/' /etc/needrestart/needrestart.conf
 fi
 
-# Install certbot - changed method for all Debian and Ubuntu versions as certbot cloudflare DNS plugin is installed from snap
-snap install core
-snap refresh core
-snap install --classic certbot
-ln -s /snap/bin/certbot /usr/bin/certbot
-if $cf; then
-    snap set certbot trust-plugin-with-root=ok
-    snap install certbot-dns-cloudflare
-fi
+if $https; then
+    if [ $(awk -F= '/^ID=/{print $2}' /etc/os-release) == "debian" ]; then
+        apt install snapd -y
+    fi
 
+    # Install certbot - changed method for all Debian and Ubuntu versions as certbot cloudflare DNS plugin is installed from snap
+    snap install core
+    snap refresh core
+    snap install --classic certbot
+    ln -s /snap/bin/certbot /usr/bin/certbot
+    if $cf; then
+        snap set certbot trust-plugin-with-root=ok
+        snap install certbot-dns-cloudflare
+    fi
+fi
 
 apt-get update
 
@@ -234,42 +238,47 @@ sed -i "s,# ssl_certificate_key /etc/letsencrypt/live/example.com/privkey.pem;,s
 ## replace remaining instances of example.com in 80/443 blocks of config file
 sed -i "s/example.com/$domain/" /etc/nginx/sites-available/mastodon
 
-## certbot if using cloudflare API token. Renewal scheduling is automatic because we don't have to use 'standalone'
-if $cf; then
-    cf_ini_dir=~/.secrets/certbot
-    mkdir -p $cf_ini_dir
-    echo '# Cloudflare API token used by Certbot' >> $cf_ini_dir/cloudflare-$domain.ini
-    echo 'dns_cloudflare_api_token = '$cloudflare_token >> $cf_ini_dir/cloudflare-$domain.ini
-    chmod 600 $cf_ini_dir/cloudflare-$domain.ini
-    
-    if $staging
-        certbot certonly --staging --dns-cloudflare --dns-cloudflare-credentials $cf_ini_dir/cloudflare-$domain.ini --dns-cloudflare-propagation-seconds 30 -d $domain --non-interactive --agree-tos -m $email
+# https is true by default, unless the user specifies 'no https' (i.e. manually configuring https)
+if $https; then
+
+    ## certbot if using cloudflare API token. Renewal scheduling is automatic because we don't have to use 'standalone'
+    if $cf; then
+        cf_ini_dir=~/.secrets/certbot
+        mkdir -p $cf_ini_dir
+        echo '# Cloudflare API token used by Certbot' >> $cf_ini_dir/cloudflare-$domain.ini
+        echo 'dns_cloudflare_api_token = '$cloudflare_token >> $cf_ini_dir/cloudflare-$domain.ini
+        chmod 600 $cf_ini_dir/cloudflare-$domain.ini
+        
+        if $staging
+            certbot certonly --staging --dns-cloudflare --dns-cloudflare-credentials $cf_ini_dir/cloudflare-$domain.ini --dns-cloudflare-propagation-seconds 30 -d $domain --non-interactive --agree-tos -m $email
+        else
+            certbot certonly --dns-cloudflare --dns-cloudflare-credentials $cf_ini_dir/cloudflare-$domain.ini --dns-cloudflare-propagation-seconds 30 -d $domain --non-interactive --agree-tos -m $email
+        fi
+
+    ## using 'standalone' certbot method. We have to take Nginx down so certbot can server on port 80 for challenge request. Renewals have to also stop/start Nginx.
     else
-        certbot certonly --dns-cloudflare --dns-cloudflare-credentials $cf_ini_dir/cloudflare-$domain.ini --dns-cloudflare-propagation-seconds 30 -d $domain --non-interactive --agree-tos -m $email
+        systemctl stop nginx
+
+        if $staging
+        then 
+            certbot certonly --staging --standalone -d $domain --non-interactive --agree-tos -m $email
+        else 
+            certbot certonly --standalone -d $domain --non-interactive --agree-tos -m $email
+        fi
+
+        systemctl start nginx
+
+        ## Certbot renewal
+        ## Very useful: https://eff-certbot.readthedocs.io/en/stable/using.html#automated-renewals
+        sudo sh -c 'printf "#!/bin/sh\nservice nginx stop\n" > /etc/letsencrypt/renewal-hooks/pre/nginx.sh'
+        sudo sh -c 'printf "#!/bin/sh\nservice nginx start\n" > /etc/letsencrypt/renewal-hooks/post/nginx.sh'
+        sudo chmod 755 /etc/letsencrypt/renewal-hooks/pre/nginx.sh
+        sudo chmod 755 /etc/letsencrypt/renewal-hooks/post/nginx.sh
+
+        SLEEPTIME=$(awk 'BEGIN{srand(); print int(rand()*(3600+1))}'); echo "0 0,12 * * * root sleep $SLEEPTIME && certbot renew -q" | tee -a /etc/crontab > /dev/null
     fi
 
-## using 'standalone' certbot method. We have to take Nginx down so certbot can server on port 80 for challenge request. Renewals have to also stop/start Nginx.
-else
-    systemctl stop nginx
-
-    if $staging
-    then 
-        certbot certonly --staging --standalone -d $domain --non-interactive --agree-tos -m $email
-    else 
-        certbot certonly --standalone -d $domain --non-interactive --agree-tos -m $email
-    fi
-
-    systemctl start nginx
-
-    ## Certbot renewal
-    ## Very useful: https://eff-certbot.readthedocs.io/en/stable/using.html#automated-renewals
-    sudo sh -c 'printf "#!/bin/sh\nservice nginx stop\n" > /etc/letsencrypt/renewal-hooks/pre/nginx.sh'
-    sudo sh -c 'printf "#!/bin/sh\nservice nginx start\n" > /etc/letsencrypt/renewal-hooks/post/nginx.sh'
-    sudo chmod 755 /etc/letsencrypt/renewal-hooks/pre/nginx.sh
-    sudo chmod 755 /etc/letsencrypt/renewal-hooks/post/nginx.sh
-
-    SLEEPTIME=$(awk 'BEGIN{srand(); print int(rand()*(3600+1))}'); echo "0 0,12 * * * root sleep $SLEEPTIME && certbot renew -q" | tee -a /etc/crontab > /dev/null
-fi
+fi # if $https end
 
 ## Copy, reload and enable mastodon services
 cp /home/mastodon/live/dist/mastodon-*.service /etc/systemd/system/
